+++ memberaudit/tasks.py
@@ -69,10 +69,15 @@ def update_all_characters(self, force_update: bool = False) -> None:
         stats = CharacterUpdateStatus.objects.statistics()
         logger.info(f"Update statistics: {stats}")
 
-    characters_with_owners = Character.objects.filter(
-        eve_character__character_ownership__isnull=False, is_disabled=False
-    ).values_list("pk", flat=True)
-    for character_pk in characters_with_owners:
+    # disable characters without owner
+    Character.objects.filter(eve_character__character_ownership__isnull=True).update(
+        is_disabled=True
+    )
+    # start sync for all enabled characters
+    enabled_characters = Character.objects.filter(is_disabled=False).values_list(
+        "pk", flat=True
+    )
+    for character_pk in enabled_characters:
         update_character.apply_async(
             kwargs={"character_pk": character_pk, "force_update": force_update},
             priority=DEFAULT_TASK_PRIORITY,