+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Filter with counts for "last update ok" on characters admin page
 - Ability to disable sync for characters, e.g. which no longer have a valid token
 - Filter with counts for state incl. filtering characters without main
+- Automatically disable characters which no longer have an owner
 
 ## [2.4.2] - 2023-01-31